requirements.txt
Django>=2.1.6
django-cors-headers==2.4.0
django-filter==2.1.0
djangorestframework==3.9.1
Markdown==3.0.1
pytz==2018.9
PyMySQL==0.9.3
whitenoise==4.1.2
